Use one-based numbering for the `retry.attempt` logger metadata value instead of zero-based numbering.
A future commit will log the maximum allowed number of attempts. The log message when using the Apple logger will have the phrase “Attempt 1 of 3 failed”, which sounds more natural with one-based numbering.

This is not a breaking change because the commit that made the logger metadata keys public has not yet been released.